+++ pom.xml
@@ -82,7 +82,7 @@
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>org.cornfields.simulator.CornfieldSimApplication</mainClass>
+                                    <mainClass>org.cornfields.simulator.CornApplication</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
