+++ src/main/java/org/cornfields/simulator/command/CommandFactory.java
@@ -1,4 +1,4 @@
-package org.cornfields.simulator.game;
+package org.cornfields.simulator.command;
 
 import java.util.Optional;
