+++ src/test/java/org/cornfields/simulator/game/CommandProcessorTest.java
@@ -1,8 +1,25 @@
 package org.cornfields.simulator.game;
 
+import org.cornfields.simulator.CommandNotAllowedException;
+import org.cornfields.simulator.command.RegisterCommand;
+import org.cornfields.simulator.command.TravelCommand;
 import org.junit.Test;
 
 public class CommandProcessorTest {
+
+  @Test
+  public void testRegisterAliasTooShort() {
+    final CommandProcessor processor = new CommandProcessor();
+
+    try {
+
+      processor.process(new RegisterCommand("123", ""));
+      assert false;
+
+    } catch (CommandNotAllowedException e) {
+      assert true;
+    }
+  }
 
   @Test
   public void testRegisterAliasTooLong() {
@@ -10,7 +27,7 @@ public class CommandProcessorTest {
 
     try {
 
-      processor.process(new RegisterCommand("123", "dogeeeee"));
+      processor.process(new RegisterCommand("123", "dogeeeeeee"));
       assert false;
 
     } catch (CommandNotAllowedException e) {
